Refactor for server side for Spring Cloud Contract
