+++ server/src/contractTest/resources/contracts/employees/shouldReturnSampleEmployee.groovy
@@ -13,10 +13,10 @@ Contract.make {
         headers {
             contentType applicationJson()
         }
-        body("""{
-            "uuid": "00000000-0000-0000-a000-000000000001",
-            "name": "John Doe",
-            "email": "john.doe@example.com"
-        }""")
+        body(
+                uuid: "00000000-0000-0000-a000-000000000001",
+                name: "John Doe",
+                email: "john.doe@example.com"
+        )
     }
 }
