+++ server/src/main/kotlin/pl/kubiczak/test/spring/integration/demo/server/employees/EmployeesController.kt
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RestController
-import java.util.*
+import java.util.UUID
 
 @RestController
 class EmployeesController(
@@ -18,8 +18,8 @@ class EmployeesController(
         )
 
     @GetMapping(path = ["/employees/{uuid}"])
-    fun readAll(@PathVariable uuid: UUID): ResponseEntity<EmployeeDto> =
-        ResponseEntity.of(
-            employeesService.findByUuid(uuid)
-        )
+    fun readAll(@PathVariable uuid: UUID): ResponseEntity<EmployeeDto> {
+        val employee = employeesService.findByUuid(uuid)
+        return ResponseEntity.of(employee)
+    }
 }
