+++ server/src/test/groovy/pl/kubiczak/test/spring/integration/demo/server/TestcontainersSpringBaseTest.groovy
@@ -1,8 +1,13 @@
 package pl.kubiczak.test.spring.integration.demo.server
 
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.boot.test.util.TestPropertyValues
 import org.springframework.boot.test.web.server.LocalServerPort
+import org.springframework.context.ApplicationContextInitializer
+import org.springframework.context.ConfigurableApplicationContext
 import org.springframework.test.context.ContextConfiguration
+import org.testcontainers.containers.PostgreSQLContainer
+import org.testcontainers.junit.jupiter.Container
 import org.testcontainers.spock.Testcontainers
 import spock.lang.Specification
 
@@ -17,5 +22,23 @@ class TestcontainersSpringBaseTest extends Specification {
 
     @LocalServerPort
     protected int localServerPort
+
+    static class TestcontainersDbInitializer implements ApplicationContextInitializer<ConfigurableApplicationContext> {
+
+        @Container
+        public static final PostgreSQLContainer CONTAINER = new PostgreSQLContainer<>("postgres:12.3")
+
+        @Override
+        void initialize(ConfigurableApplicationContext context) {
+            CONTAINER.start()
+            TestPropertyValues
+                    .of(
+                            "spring.datasource.url=" + CONTAINER.getJdbcUrl(),
+                            "spring.datasource.username=" + CONTAINER.getUsername(),
+                            "spring.datasource.password=" + CONTAINER.getPassword()
+                    )
+                    .applyTo(context.getEnvironment())
+        }
+    }
 
 }
