+++ server/src/test/groovy/pl/kubiczak/test/spring/integration/demo/server/employees/jdbc/EmployeeRepositoryDataTestSpec.groovy
@@ -9,7 +9,7 @@ import spock.lang.Ignore
 import spock.lang.Specification
 
 @DataJdbcTest
-class EmployeeRepositoryDataJdbcTestSpec extends Specification {
+class EmployeeRepositoryDataTestSpec extends Specification {
 
     @Autowired
     NamedParameterJdbcTemplate namedParameterJdbcTemplate
